kész

+++ VPN Connection/VPN Connection/vpn.cs
@@ -15,71 +15,80 @@ namespace VPN_Connection {
         public string error { get; set; }
 
         private RasPhoneBook createPhoneBook() {
-            string ip = resolveIP(vpnData.host).ToString();
+            string ip;
+            if ((ip = resolveIP(vpnData.host).ToString()) == null){
+                logging.writeToLog(null, String.Format("[createPhoneBook] {0} not found",vpnData.host));
+                return null;
+            }
             if(File.Exists(path)) {
                 try {
                     System.IO.File.Delete(path);
-                    logging.writeToLog(null, String.Format("[ConnectToPPTP][Remove old phonebook] success",3));
+                    logging.writeToLog(null, String.Format("[createPhoneBook][Remove old phonebook] success", 3));
                 }
                 catch(Exception e) {
-                    logging.writeToLog(null, String.Format("[ConnectToPPTP][Remove old phonebook] Exception found: {0}", e.Message));
+                    logging.writeToLog(null, String.Format("[createPhoneBook][Remove old phonebook] Exception found: {0}", e.Message));
                     return null;
                 }
             }
             try {
                 using(FileStream fs = File.Create(path)) {
                     fs.Close();
-                    logging.writeToLog(null, String.Format("[ConnectToPPTP][Create Phonebook] success",2));
+                    logging.writeToLog(null, String.Format("[createPhoneBook] success", 2));
                 }
             }
             catch(Exception e) {
-                logging.writeToLog(null, String.Format("[ConnectToPPTP][Create Phonebook] Exception found: {0}", e.Message));
+                logging.writeToLog(null, String.Format("[createPhoneBook] Exception found: {0}", e.Message));
                 return null;
             }
 
             RasPhoneBook book = new RasPhoneBook();
             try {
-                book.Open(path); //Define book path
-                logging.writeToLog(null, String.Format("[ConnectToPPTP][Open Phonebook] Success"),2);
+                book.Open(path);
+                logging.writeToLog(null, String.Format("[createPhoneBook][Open Phonebook] Success"),2);
             }
             catch(Exception e) {
-                logging.writeToLog(null, String.Format("[ConnectToPPTP][Open Phonebook] Exception found: {0}", e.Message));
+                logging.writeToLog(null, String.Format("[createPhoneBook][Open Phonebook] Exception found: {0}", e.Message));
                 return null;
             }
             RasDevice device = RasDevice.GetDevices().Where(o => o.Name.Contains("PPTP")).FirstOrDefault();
             if(device == null) {
-                logging.writeToLog(null, String.Format("[ConnectToPPTP][Device] Useable device for VPN not found"),1);
+                logging.writeToLog(null, String.Format("[createPhoneBook][Device] Useable device for VPN not found"),1);
 
             }
             RasEntry entry = RasEntry.CreateVpnEntry(vpnData.entryName, ip, RasVpnStrategy.PptpFirst, device);
             try {
                 book.Entries.Add(entry);
-                logging.writeToLog(null, String.Format("[ConnectToPPTP][Write to Phonebook] Success"),2);
+                logging.writeToLog(null, String.Format("[createPhoneBook][Write to Phonebook] Success"),2);
             }
             catch(Exception e) {
-                logging.writeToLog(null, String.Format("[ConnectToPPTP][Add phonebook entry] Exception found: {0}", e.Message));
+                logging.writeToLog(null, String.Format("[createPhoneBook][Add phonebook entry] Exception found: {0}", e.Message));
                 return null;
             }
             return book;
         }
 
         public void Dialer() {
-            logging.writeToLog(null, String.Format("Dialer] Begin"),3);
+            logging.writeToLog(null, String.Format("[Dialer] Begin"),3);
 
             RasPhoneBook book;
             if((book = createPhoneBook()) != null) {
                 try {
                     string ip = resolveIP(vpnData.host).ToString();
-                RasDialer dialer = new RasDialer();
-                dialer.PhoneBookPath = book.Path;
-                dialer.Credentials = new NetworkCredential(vpnData.username, vpnData.password);
-                dialer.EntryName = vpnData.entryName;
-                
-                    dialer.Dial();        
-                    //dialer.DialAsync();            
-                    logging.writeToLog(null, String.Format("[Dialer] Success"),2);
+                    RasDialer dialer = new RasDialer();
+                    dialer.PhoneBookPath = book.Path;
+                    dialer.Credentials = new NetworkCredential(vpnData.username, vpnData.password);
+                    dialer.EntryName = vpnData.entryName;
+                    dialer.DialCompleted += (sender, args) => {
+                        logging.writeToLog(null, String.Format("[Dialer] DialCompleted"), 2);
+                    };
+                    dialer.Error += (sender, args) => {
+                        logging.writeToLog(null, String.Format("[Dialer] DialError"), 2);
+                    };
+                    //dialer.DialAsync();
+                    dialer.Dial();
+                    logging.writeToLog(null, String.Format("[Dialer] Success"), 2);
                 }
-                catch(Exception e) {
+                catch (Exception e) {
                     logging.writeToLog(null, String.Format("[Dialer][Exception] {0}", e.Message));
                 }
             }
@@ -93,7 +102,7 @@ namespace VPN_Connection {
                 RasConnection conn;
                 if((conn = getConnectionStatus()) != null) {
                     conn.HangUp(true);
-                    logging.writeToLog(null, String.Format("[disconnectPPTP] Disconnect Success"), 2);
+                    logging.writeToLog(null, String.Format("[disconnectPPTP] Disconnect Success"), 1);
                 }
                 conn = null;
             };
@@ -116,10 +125,15 @@ namespace VPN_Connection {
         public IPAddress resolveIP(string host) {
             try {
                 using(Ping Ping = new Ping()) {
-                    PingReply PingReply = Ping.Send(host);
-                    if(PingReply.Status == IPStatus.Success) {
-                        logging.writeToLog(null, String.Format("[resolveIP] {0} found, IP address: {1}", host, PingReply.Address), 2);
-                        return PingReply.Address;
+                    try {
+                        PingReply PingReply = Ping.Send(host);
+                        if (PingReply.Status == IPStatus.Success) {
+                            logging.writeToLog(null, String.Format("[resolveIP] {0} found, IP address: {1}", host, PingReply.Address), 2);
+                            return PingReply.Address;
+                        }
+                    }
+                    catch(Exception e) {
+                        logging.writeToLog(null, String.Format("[resolveIP][Exception] Ping {1} thrown an exception: ", host, e.Message));
                     }
                 }
             }
